requirements.txt
asgiref==3.4.1
Django==3.2.3
django-bulma==0.8.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
dnspython==2.1.0
gunicorn==20.0.4
PyJWT==2.1.0
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
typing-extensions==3.10.0.1
uritemplate==3.0.1
psycopg2-binary==2.8.6
dj-database-url==0.5.0
django-heroku==0.3.1